Split create_minion
A node is just a server created by the provider. A minion is a node
that is accepted on the saltmaster ("connected"). Thus the full provider
actions for nodes and minions are as follows:

provider -> create -> node
node -> connect -> minion
minion -> disconnect -> node
node -> destroy

Create/destroy are thus node-level operations, while connect/disconnect
are minion operations. Creating or destroying a minion is a shortcut
to both create and connect or disconnect and destroy.